# app.js
/************************************************************************
 * Objetivo:
 ***********************************************************************/
//Import das dependencias  para criar a API
const express = require('express')
const cors = require('cors')
const bodyParser = require('body-parser')

//Import das CONTROLLERS do projeto
const controllerFilme = require('./controller/filme/controller_filme.js')
const controllerSexo = require('./controller/sexo/controller_sexo.js')
const controllerClassificacao = require('./controller/classificacao/controller_classificacao.js')
const controllerGenero = require('./controller/genero/controller_genero.js')

//Criando um objeto para manipular dados do body da API em formato JSON
const bodyParserJson = bodyParser.json()

//Criando um objeto para manipular o express
const app = express()

//Conjunto de permissões a serem aplicadas no CORS da API 
const corsOptions = {
    origin: ['*'], // A origem da requisição, podendos ser um ip ou *(todos)
    methods: 'GET, POST, PUT, DELETE, OPTIONS', // São os verbos que seram liberados na API (GET, POST, PUT E DELETE)
    allowedHeaders: ['Content-type', 'Autorization'], // São permissões de cabeçalho do CORS
}

//Configura as permissões da API através do CORS
app.use(cors(corsOptions))

//ENDPOINTS FILME
app.post('/v1/senai/locadora/filme', bodyParserJson, async function (request, response) {
    //Recebe o conteúdo dentro do body da requisição
    let dados = request.body
    //Recebe o content type da requisição, para validar se é um JSON
    let contentType = request.headers['content-type']

    let result = await controllerFilme.inserirNovoFilme(dados, contentType)

    response.status(result.status_code)
    response.json(result)
})

app.get('/v1/senai/locadora/filme', async function (request, response) {
    let result = await controllerFilme.listarFilme()

    response.status(result.status_code)
    response.json(result)

})

app.get('/v1/senai/locadora/filme/:id', async function (request, response) {
    let id = request.params.id

    let result = await controllerFilme.buscarFilme(id)

    response.status(result.status_code)
    response.json(result)
})

app.put('/v1/senai/locadora/filme/:id', bodyParserJson, async function (request, response) {
    //Recebe o content type da requisição
    let contentType = request.headers['content-type']

    //Recebe o id do registro a ser atualizado
    let id = request.params.id

    //Recebe os dados enviados no corpo da requisição
    let dados = request.body

    //Chama a função de atualizar na controller e encaminha os dados, id e content-type
    //obedecendo a ordem de criação na função da controller
    let result = await controllerFilme.atualizarFilme(dados, id, contentType)


    response.status(result.status_code)
    response.json(result)
})

app.delete('/v1/senai/locadora/filme/:id', async function (request, response) {
    let id = request.params.id

    let result = await controllerFilme.excluirFilme(id)

    response.status(result.status_code)
    response.json(result)
})

//ENDPOINTS SEXO
app.post('/v1/senai/locadora/sexo', bodyParserJson, async function (request, response) {
    //Recebe o conteúdo dentro do body da requisição
    let dados = request.body
    //Recebe o content type da requisição, para validar se é um JSON
    let contentType = request.headers['content-type']

    let result = await controllerSexo.inserirSexo(dados, contentType)

    response.status(result.status_code)
    response.json(result)
})

app.get('/v1/senai/locadora/sexo', async function (request, response) {
    let result = await controllerSexo.listarSexo()

    response.status(result.status_code)
    response.json(result)

})

app.get('/v1/senai/locadora/sexo/:id', async function (request, response) {
    let id = request.params.id

    let result = await controllerSexo.buscarSexo(id)

    response.status(result.status_code)
    response.json(result)
})

app.put('/v1/senai/locadora/sexo/:id', bodyParserJson, async function (request, response) {
    //Recebe o content type da requisição
    let contentType = request.headers['content-type']

    //Recebe o id do registro a ser atualizado
    let id = request.params.id

    //Recebe os dados enviados no corpo da requisição
    let dados = request.body

    //Chama a função de atualizar na controller e encaminha os dados, id e content-type
    //obedecendo a ordem de criação na função da controller
    let result = await controllerSexo.atualizarSexo(dados, id, contentType)


    response.status(result.status_code)
    response.json(result)
})

app.delete('/v1/senai/locadora/sexo/:id', async function (request, response) {
    let id = request.params.id

    let result = await controllerSexo.excluirSexo(id)

    response.status(result.status_code)
    response.json(result)
})

//CLASSIFICACAO
app.post('/v1/senai/locadora/classificacao', bodyParserJson, async function (request, response) {
    //Recebe o conteúdo dentro do body da requisição
    let dados = request.body

    //Recebe o content type da requisição, para validar se é um JSON
    let contentType = request.headers['content-type']

    let result = await controllerClassificacao.inserirNovoClassificacao(dados, contentType)
    console.log(result)
    response.status(result.status_code)
    response.json(result)
})

app.get('/v1/senai/locadora/classificacao', async function (request, response) {
    let result = await controllerClassificacao.listarClassificacao()

    response.status(result.status_code)
    response.json(result)
})

app.get('/v1/senai/locadora/classificacao/:id', async function (request, response) {
    //Recebe o ID via parametro
    let id = request.params.id

    let result = await controllerClassificacao.buscarClassificacao(id)

    response.status(result.status_code)
    response.json(result)
})

app.put('/v1/senai/locadora/classificacao/:id', bodyParserJson, async function (request, response) {
    //Recebe o contenty type da requisição
    let contentType = request.headers['content-type']
    //Recebe o ID do registro a ser atualizado
    let id = request.params.id
    //Recebe os dados enviados no corpo da requisição
    let dados = request.body

    //Chama a função de atualizar na controller e encaminha os dados, id e content-type
    //obedecendo a ordem de criação na função da controller
    let result = await controllerClassificacao.atualizarClassificacao(dados, id, contentType)

    response.status(result.status_code)
    response.json(result)
})

app.delete('/v1/senai/locadora/classificacao/:id', async function (request, response) {
    let id = request.params.id

    let result = await controllerClassificacao.excluirclassificacao(id)

    response.status(result.status_code)
    response.json(result)
})

//GENERO
app.post('/v1/senai/locadora/genero', bodyParserJson, async function (request, response) {

    //recebe o conteúdo
    let dados = request.body
    let contentType = request.headers['content-type'] //recebe o content type da requisição para validar se é um json

    let result = await controllerGenero.inserirNovoGenero(dados, contentType)


    response.status(result.status_code)
    response.json(result)
})

app.get('/v1/senai/locadora/genero', async function (request, response) {
    let result = await controllerGenero.listarGenero()

    response.status(result.status_code)
    response.json(result)
})

app.get('/v1/senai/locadora/genero/:id', async function (request, response) {
    let id = request.params.id

    let result = await controllerGenero.buscarGenero(id)

    response.status(result.status_code)
    response.json(result)
})

app.put('/v1/senai/locadora/genero/:id', bodyParserJson, async function (request, response) {

    //Recebe o contenty type da requisição
    let contentType = request.headers['content-type']

    //Receber o ID do registro a ser atulizado
    let id = request.params.id

    //Receber os dados enviados no corpo de requisição
    let dados = request.body

    //Chama a função de atualizar na controller e encaminhando os dados, id e content-type
    //obedecendo a ordem de criação na função da controller
    let result = await controllerGenero.atualizarGenero(dados, id, contentType)

    response.status(result.status_code)
    response.json(result)

})

app.delete('/v1/senai/locadora/genero/:id', async function (request, response) {


    let id = request.params.id

    let result = await controllerGenero.excluirGenero(id)

    response.status(result.status_code)
    response.json(result)

})


//Serve para inicializar a API para receber requisições 
app.listen(8080, function () {
    console.log('Api funcionando e aguardando novas requisições ...')
})